+++ concordium-sdk/src/main/java/com/concordium/sdk/exceptions/CryptoJniException.java
@@ -34,4 +34,3 @@ public final class CryptoJniException extends RuntimeException {
         return new CryptoJniException(code);
     }
 }
-