+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/TransactionType.java
@@ -22,7 +22,10 @@ public enum TransactionType {
     REGISTER_DATA((byte) 21),
     TRANSFER_WITH_MEMO((byte) 22),
     ENCRYPTED_TRANSFER_WITH_MEMO((byte) 23),
-    TRANSFER_WITH_SCHEDULE_AND_MEMO((byte) 24);
+    TRANSFER_WITH_SCHEDULE_AND_MEMO((byte) 24),
+    CONFIGURE_BAKER((byte) 25),
+    CONFIGURE_DELEGATION((byte) 26);
+
 
     /**
      * Number of Bytes used for Serializing {@link TransactionType}.