+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased changes
+
+## 4.1.0
 - Added support for Transfer With Schedule.
 - Added support for Transfer With Schedule With Memo.
 - Added support for Update Credentials Keys.
@@ -8,6 +10,10 @@
 - Added support for Transfer To Encrypted wallet.
 - Added support for Encrypted Transfer.
 - Added support for Encrypted Transfer with Memo.
+- Added support for Configuring a baker.
+- Added support for Configuring the account as a Delegator.
+- A new native dependency has been introduced in order to support the new transaction types mentioned above,
+  As a result native dependencies need to be rebuilt before building the Java project.
 - Added initial support for GRPC V2
 - Added initial support for GRPC V2 query GetAnonymityRevokers
 - Added CLI example for GRPC V2 query GetAnonymityRevokers